feat: create form data to store obat racikan and handling some error in authorization

+++ app/Database/Migrations/2024-05-07-174748_Rekmed.php
@@ -75,7 +75,7 @@ class Rekmed extends Migration
                 'null' => true
             ],
             'imt' => [
-                'type' => 'INT',
+                'type' => 'FLOAT',
                 'constraint' => 11,
                 'null' => true
             ],
